+++ R/SBD.R
@@ -12,6 +12,17 @@
 #'
 #' The output values lie between 0 and 2, with 0 indicating perfect similarity.
 #'
+#' @examples
+#'
+#' # load data
+#' data(uciCT)
+#'
+#' # distance between series of different lengths
+#' sbd <- SBD(CharTraj[[1]], CharTraj[[100]], znorm = TRUE)$dist
+#'
+#' # cross-distance matrix for series subset (notice the two-list input)
+#' sbD <- proxy::dist(CharTraj[1:10], CharTraj[1:10], method = "SBD", znorm = TRUE)
+#'
 #' @seealso
 #'
 #' \code{\link{NCCc}}, \code{\link{shape_extraction}}